refactor: move Build and BuildAndPublish functions to dedicated file
